Convert platforms to live data
Required changing order of colors so they would match up with the
correct platforms, which is a bad fix. Colors for platforms should be
stored in an object so the colors can be mapped to platform names, but
this will require a rewrite of the pie chart element's color handling.